Renamed the addImage use case to editImages
