+++ src/main/java/interface_adapter/editImages/EditImagesPresenter.java
@@ -1,4 +1,4 @@
-package interface_adapter.addImage;
+package interface_adapter.editImages;
 
 import use_case.editImages.EditImagesOutputBoundary;
 import use_case.editImages.EditImagesOutputData;
