+++ src/main/java/interface_adapter/editImages/EditImagesViewModel.java
@@ -1,4 +1,4 @@
-package interface_adapter.addImage;
+package interface_adapter.editImages;
 
 import interface_adapter.ViewModel;
